Added Collectibles + Score with Scripts.
