+++ Assets/Animations/Key.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3acf60e422d74644c981707cd37e3f73
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
