+++ Assets/Animations/Key/Key.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 16f052ccff295fc4bbbe84f4231c731d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
