+++ Assets/Animations/Key/Key_Spin.anim
@@ -1,0 +1,155 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!74 &7400000
+AnimationClip:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: Key_Spin
+  serializedVersion: 6
+  m_Legacy: 0
+  m_Compressed: 0
+  m_UseHighQualityCurve: 1
+  m_RotationCurves: []
+  m_CompressedRotationCurves: []
+  m_EulerCurves: []
+  m_PositionCurves: []
+  m_ScaleCurves: []
+  m_FloatCurves: []
+  m_PPtrCurves:
+  - curve:
+    - time: 0
+      value: {fileID: 21300000, guid: 3abf1705bc9d76244ad760412db78a65, type: 3}
+    - time: 0.033333335
+      value: {fileID: 21300000, guid: f16737ca9f1dc29448765a816d0a1b49, type: 3}
+    - time: 0.06666667
+      value: {fileID: 21300000, guid: 7a0de91dea9f59140bb693004bf0b0ad, type: 3}
+    - time: 0.1
+      value: {fileID: 21300000, guid: 7c8bdf6129a84e141a690926a43cb384, type: 3}
+    - time: 0.13333334
+      value: {fileID: 21300000, guid: 2a8146b76a8332e46b9c00f3cb2960e8, type: 3}
+    - time: 0.16666667
+      value: {fileID: 21300000, guid: 2c9d2d33aa1cb054cb019c852a64b6d4, type: 3}
+    - time: 0.2
+      value: {fileID: 21300000, guid: 6b805b3d02c25084780e70f6b6e7ab38, type: 3}
+    - time: 0.23333333
+      value: {fileID: 21300000, guid: b53184e6a2932ee409a26e6838f16aae, type: 3}
+    - time: 0.28333333
+      value: {fileID: 21300000, guid: 44fbfa5afc056fe459c80663e42e9ae4, type: 3}
+    - time: 0.31666666
+      value: {fileID: 21300000, guid: 862eb60c627dd534cb832145605ac12d, type: 3}
+    - time: 0.35
+      value: {fileID: 21300000, guid: 7a85663a637e4004e92472acad01c2b8, type: 3}
+    - time: 0.38333333
+      value: {fileID: 21300000, guid: e800ef09025e509499b2916a88d7da82, type: 3}
+    - time: 0.41666666
+      value: {fileID: 21300000, guid: 3dbc35981e1cf7b4e8e17794cc775b4d, type: 3}
+    - time: 0.45
+      value: {fileID: 21300000, guid: 0a3bfa4764855014695c9c734bd910a5, type: 3}
+    - time: 0.48333332
+      value: {fileID: 21300000, guid: 2c458ee8cd0fbbf4fa714059d5f3136d, type: 3}
+    - time: 0.51666665
+      value: {fileID: 21300000, guid: b940cb8a87dbc2b4c9924fa949455cab, type: 3}
+    - time: 0.55
+      value: {fileID: 21300000, guid: 36d7a8d875f6bec499f358447156b5fc, type: 3}
+    - time: 0.5833333
+      value: {fileID: 21300000, guid: 5bc4e6fd2df715c46b166367924563d2, type: 3}
+    - time: 0.6166667
+      value: {fileID: 21300000, guid: c7e105af68f080d408ff8dcfc71d976f, type: 3}
+    - time: 0.65
+      value: {fileID: 21300000, guid: 55b7d14da52e94f40a7977ab1fd4678e, type: 3}
+    - time: 0.68333334
+      value: {fileID: 21300000, guid: 75c23d1c7c70bb847b9946964aaf0825, type: 3}
+    - time: 0.71666664
+      value: {fileID: 21300000, guid: cd5d447894330194aa31e955895c2404, type: 3}
+    - time: 0.76666665
+      value: {fileID: 21300000, guid: 854bacb421a0cf0409b811ca82dc0b72, type: 3}
+    - time: 0.8
+      value: {fileID: 21300000, guid: 36cb6f5bee0e9c64b8fd1b8d8fa3a779, type: 3}
+    - time: 0.8333333
+      value: {fileID: 21300000, guid: e2eadf32de58de044a6101c5d48fb7d5, type: 3}
+    - time: 0.8666667
+      value: {fileID: 21300000, guid: 2e67ec8e00de06747930191c34bec57b, type: 3}
+    - time: 0.9
+      value: {fileID: 21300000, guid: d48d2bef2814aec4db2752ccf322c364, type: 3}
+    - time: 0.93333334
+      value: {fileID: 21300000, guid: ee67da4db316077448ea9b6e1f5429b4, type: 3}
+    - time: 0.96666664
+      value: {fileID: 21300000, guid: 3ece3017d2ee76f4d84987cb7e02ea37, type: 3}
+    - time: 1
+      value: {fileID: 21300000, guid: a74e1b80bec2c7a41b68c0a4e6a2085a, type: 3}
+    attribute: m_Sprite
+    path: 
+    classID: 212
+    script: {fileID: 0}
+  m_SampleRate: 60
+  m_WrapMode: 0
+  m_Bounds:
+    m_Center: {x: 0, y: 0, z: 0}
+    m_Extent: {x: 0, y: 0, z: 0}
+  m_ClipBindingConstant:
+    genericBindings:
+    - serializedVersion: 2
+      path: 0
+      attribute: 0
+      script: {fileID: 0}
+      typeID: 212
+      customType: 23
+      isPPtrCurve: 1
+    pptrCurveMapping:
+    - {fileID: 21300000, guid: 3abf1705bc9d76244ad760412db78a65, type: 3}
+    - {fileID: 21300000, guid: f16737ca9f1dc29448765a816d0a1b49, type: 3}
+    - {fileID: 21300000, guid: 7a0de91dea9f59140bb693004bf0b0ad, type: 3}
+    - {fileID: 21300000, guid: 7c8bdf6129a84e141a690926a43cb384, type: 3}
+    - {fileID: 21300000, guid: 2a8146b76a8332e46b9c00f3cb2960e8, type: 3}
+    - {fileID: 21300000, guid: 2c9d2d33aa1cb054cb019c852a64b6d4, type: 3}
+    - {fileID: 21300000, guid: 6b805b3d02c25084780e70f6b6e7ab38, type: 3}
+    - {fileID: 21300000, guid: b53184e6a2932ee409a26e6838f16aae, type: 3}
+    - {fileID: 21300000, guid: 44fbfa5afc056fe459c80663e42e9ae4, type: 3}
+    - {fileID: 21300000, guid: 862eb60c627dd534cb832145605ac12d, type: 3}
+    - {fileID: 21300000, guid: 7a85663a637e4004e92472acad01c2b8, type: 3}
+    - {fileID: 21300000, guid: e800ef09025e509499b2916a88d7da82, type: 3}
+    - {fileID: 21300000, guid: 3dbc35981e1cf7b4e8e17794cc775b4d, type: 3}
+    - {fileID: 21300000, guid: 0a3bfa4764855014695c9c734bd910a5, type: 3}
+    - {fileID: 21300000, guid: 2c458ee8cd0fbbf4fa714059d5f3136d, type: 3}
+    - {fileID: 21300000, guid: b940cb8a87dbc2b4c9924fa949455cab, type: 3}
+    - {fileID: 21300000, guid: 36d7a8d875f6bec499f358447156b5fc, type: 3}
+    - {fileID: 21300000, guid: 5bc4e6fd2df715c46b166367924563d2, type: 3}
+    - {fileID: 21300000, guid: c7e105af68f080d408ff8dcfc71d976f, type: 3}
+    - {fileID: 21300000, guid: 55b7d14da52e94f40a7977ab1fd4678e, type: 3}
+    - {fileID: 21300000, guid: 75c23d1c7c70bb847b9946964aaf0825, type: 3}
+    - {fileID: 21300000, guid: cd5d447894330194aa31e955895c2404, type: 3}
+    - {fileID: 21300000, guid: 854bacb421a0cf0409b811ca82dc0b72, type: 3}
+    - {fileID: 21300000, guid: 36cb6f5bee0e9c64b8fd1b8d8fa3a779, type: 3}
+    - {fileID: 21300000, guid: e2eadf32de58de044a6101c5d48fb7d5, type: 3}
+    - {fileID: 21300000, guid: 2e67ec8e00de06747930191c34bec57b, type: 3}
+    - {fileID: 21300000, guid: d48d2bef2814aec4db2752ccf322c364, type: 3}
+    - {fileID: 21300000, guid: ee67da4db316077448ea9b6e1f5429b4, type: 3}
+    - {fileID: 21300000, guid: 3ece3017d2ee76f4d84987cb7e02ea37, type: 3}
+    - {fileID: 21300000, guid: a74e1b80bec2c7a41b68c0a4e6a2085a, type: 3}
+  m_AnimationClipSettings:
+    serializedVersion: 2
+    m_AdditiveReferencePoseClip: {fileID: 0}
+    m_AdditiveReferencePoseTime: 0
+    m_StartTime: 0
+    m_StopTime: 1.0166667
+    m_OrientationOffsetY: 0
+    m_Level: 0
+    m_CycleOffset: 0
+    m_HasAdditiveReferencePose: 0
+    m_LoopTime: 1
+    m_LoopBlend: 0
+    m_LoopBlendOrientation: 0
+    m_LoopBlendPositionY: 0
+    m_LoopBlendPositionXZ: 0
+    m_KeepOriginalOrientation: 0
+    m_KeepOriginalPositionY: 1
+    m_KeepOriginalPositionXZ: 0
+    m_HeightFromFeet: 0
+    m_Mirror: 0
+  m_EditorCurves: []
+  m_EulerEditorCurves: []
+  m_HasGenericRootTransform: 0
+  m_HasMotionFloatCurves: 0
+  m_Events: []
