+++ Assets/Animations/Key/Key_Spin.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c18a27aede004d840b1576ebb57db07a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
