+++ Assets/Scripts/KeyController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f12843d7f4d84954baf970d8e761047a
+guid: 0d4487672f1d1c3478d97dff8a810472
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
