+++ Assets/Scripts/ScoreController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7b7ac91895e804f43b7532765bcfb73a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
